backend/requirements.txt
anyio==4.3.0
asgiref==3.8.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Django==5.0.4
django-admin-interface==0.28.6
django-admin-tools==0.9.3
django-appconf==1.0.6
django-colorfield==0.11.0
django-imagekit==5.0.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
pilkit==3.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-slugify==8.0.4
sniffio==1.3.1
sqlparse==0.5.0
starlette==0.37.2
text-unidecode==1.3
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
